Virtual assignment (#50858)
This makes compiler dependencies much easier to read and understand.

## Motivation

This also allows users to see directly, in the output of `spack spec`
and other commands, what compiler was used for a node. It can be extended
later to show things like which MPI or other virtual was chosen for a
given node in `spack spec` output.  This essentially restores `spack spec`
output to where it was in 0.23, and will:

  1. Allow users to see what compilers were chosen
  2. Allow users to understand how to specify them. 

`%c=gcc`, vs.` %[virtuals=c] gcc` may not seem like much, but consider for
mixed compilers that you are likely to be writing:

```
%[virtuals=c,cxx] gcc %[virtuals=fortran] gfortran
```

vs.:

```
%c= gcc %fortran=gfortran
```

The latter is easy enough to specify ad-hoc on the CLI and is much easier
to remember and read.  Consider that users are likely to glance at this as
part of `spack spec` output -- the former is very noisy in that context.
4. Allow users to specify toolchains more easily.

Without this, users are going to need to understand our edge attribute
syntax, which is not something I think should be in day-to-day use. The
intuition with the syntax here is that, and virtuals are interfaces where
you choose an implementation to substitute.  Virtuals are also now a
*very* common part of users' conception of packages in Spack, since we use
them for languages: `c`, `cxx`, and `fortran`, and in others in the future.

## Changes
Three main changes:

1. You can now write virtuals on edges as, e.g.:
    ```
    %c=gcc
    %c,cxx=gcc
    ^mpi=mpich@3.1
    ```
    instead of:
    ```
    %[virtuals=c] gcc
    %[virtuals=c,cxx] gcc
    ^[virtuals=mpi] mpich
    ```
5. `spack spec`, `spack solve`, and other places that would've shown
`%compiler@version` before will show the `c`, `cxx`, and `fortran`
virtuals for specs. This can be expanded later to be more customizable
(e.g. if you want to see the MPI on each node in `spack spec`) but for
now it allows users to easily see what compiler each node is built with,
as they did before compiler dependencies.

Example of (2):

> spack spec zlib
 -   zlib@1.3.1+optimize+pic+shared build_system=makefile arch=darwin-sequoia-m1 %c,cxx=apple-clang@17.0.0
[e]      ^apple-clang@17.0.0 build_system=bundle arch=darwin-sequoia-m1
[+]      ^compiler-wrapper@1.0 build_system=generic arch=darwin-sequoia-m1
[+]      ^gmake@4.4.1~guile build_system=generic arch=darwin-sequoia-m1 %c=apple-clang@16.0.0
[e]          ^apple-clang@16.0.0 build_system=bundle arch=darwin-sequoia-m1
[+]          ^compiler-wrapper@1.0 build_system=generic arch=darwin-sequoia-m1

## Breaking change

Previously, `^c=gcc` would've been "any spec that depends on something
that has gcc as the value of the c variant". While this is an uncommon
query, that meaning is still possible, but requires a wildcard, like
this:

```
spack find "^* foo=bar"
```

A key value pair assignment immediately after a dependency or after an
edge specifier, where the spec name would normally be, is considered
virtual assignment. Examples:

```
^c=gcc
%c=gcc
%[when=%c] c=gcc
```

Note that because we support a wildcard, you can now parse `Spec()`
on the CLI, where previously that wasn't possible. `Spec()` is just `*`.

- [x] parse virtual assignment syntax
- [x] output virtual assignment syntax in `str(spec)`, `long_spec`, etc.
- [x] print virtual assignment syntax in `DISPLAY_FORMAT` for compilers
- [x] add "{compilers}" spec format property for this
- [x] rework tests to account for new default output format

---------

Signed-off-by: Todd Gamblin <tgamblin@llnl.gov>