bugfix: ensure that terminal settings are preserved after install (#50913)
Spack uses a `llnl.util.tty.log.keyboard_input()` context manager to enable users to interact with the I/O daemon during an install. Currently, the I/O daemon just accepts `v` to toggle input and output, but we may update it in the future to allow more control of output from build processes.

The `keyboard_input()` context manager reads and stores `tty` settings on `__enter__` and restores them on `__exit__`. This was fine when Spack spawned one build process at a time, but now that they may be parallel, you can get cases like:

1. P1 starts, reads and stores term settings
2. P1 disables `ICANON` and `ECHO`
3. P2 starts, reads modified term settings
4. P1 finishes, restores original term settings
5. P2 finishes, restores wrong settings from (2) with `ICANON` and `ECHO` disabled

This has happened frequently since #47590 was merged. It was particularly noticeable to `bash` users. While `zsh` apparently makes an effort to preserve terminal settings when commands exist, `bash` does not, and users will see echo disabled after a parallel spack install.

This PR fixes the issue by pulling a `preserve_terminal_settings()` context manager out of `keyboard_input` and allowing the main Spack process to call it. This means that o matter how poorly our build processes behave, the terminal will still be reset to whatever it was before install.

This is a temporary fix in that it puts a wrapper around all build logic. We may be able to get rid of keyboard management in subprocesses now that this is done at the Spack level.